+++ packages/nova/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @cbnventures/nova
 
+## 0.15.4 - 2026-04-19
+
+No changes
+
 ## 0.15.3 - 2026-04-19
 
 ### UPDATED